:integration/test_kontrol_cse.py: init
